Remove mention of config.mk
I flatly disagree that `include config.mk` is good practice.  The
whole Makefile is configuration, and in a one-Makefile project, this
inclusion adds complication without adding clarity.  With it included,
if I want to know what $(COUNT_EXE) is (because I'm somehow confused),
I now have to look in a separate file; I've lost the property that
‘the Makefile contains all the information about how to build the
project’.  The fact that the 'Functions' segment ends with ‘Remember,
the config.mk file contains...’ rather confesses this confusion.

If (and only if) I wanted to have the same configuration in multiple
Makefiles, then it might be useful to `include config.mk`, but
(a) that's a significantly more complicated project, and
(b) as soon as the project grew to require two Makefiles, I'd start
using `autoconf` or `cmake` or something to handle _that_ situation.